Split utils
- Code folding only goes so far. Splitting the modular functions out
  from the CLI definition.
- Refactored addSecret, rmSecret, to be exported
- Bump tests to follow suit
