+++ main.go
@@ -1,14 +1,10 @@
 package main
 
 import (
-	"bufio"
 	"bytes"
-	"context"
-	"encoding/json"
 	"fmt"
 	"github.com/google/shlex"
 	"github.com/urfave/cli"
-	"io"
 	"io/ioutil"
 	"log"
 	"os"
@@ -16,165 +12,8 @@ import (
 	"strings"
 	"time"
 
-	cloudkms "cloud.google.com/go/kms/apiv1"
-	b64 "encoding/base64"
-	kmspb "google.golang.org/genproto/googleapis/cloud/kms/v1"
 	exec "os/exec"
 )
-
-// Serialized secret
-// { "name": "A_SECRET", "cypher": "0xD34DB33F", "created": "2019-05-01 13:01:27.189242799 -0500 CDT m=+0.000075907"}
-type Secret struct {
-	Name       string    `json:"name"`
-	Cyphertext string    `json:"cypher"`
-	Created    time.Time `json:"created"`
-	Encoding   string    `json:"encoding"`
-}
-
-func addSecret(to_add Secret) {
-	// Adds or Updates a secret
-	var secrets = ReadSecrets()
-	for index, element := range secrets {
-		if element.Name == to_add.Name {
-			log.Printf("Rotating entry %s", element.Name)
-			secrets[index] = secrets[len(secrets)-1] // copy last element to index i
-			secrets[len(secrets)-1] = Secret{}       // erase last element (zero value)
-			secrets = secrets[:len(secrets)-1]       // truncate slice
-		}
-	}
-	secrets = append(secrets, to_add)
-	WriteSecrets(secrets)
-}
-
-func rmSecret(secret_name string) {
-	// Remove a secret from the scuttle.json
-	var secrets = ReadSecrets()
-	for index, element := range secrets {
-		if element.Name == secret_name {
-			log.Printf("Removing entry %s", element.Name)
-			secrets[index] = secrets[len(secrets)-1] // copy last element to index i
-			secrets[len(secrets)-1] = Secret{}       // erase last element (zero value)
-			secrets = secrets[:len(secrets)-1]       // truncate slice
-		}
-	}
-	WriteSecrets(secrets)
-}
-
-func ReadSecrets() []Secret {
-	file, err := ioutil.ReadFile(".scuttle.json")
-	if err != nil {
-		return []Secret{}
-	}
-
-	var data []Secret
-	err = json.Unmarshal([]byte(file), &data)
-	if err != nil {
-		log.Fatal(err)
-	}
-	return data
-}
-
-func WriteSecrets(data []Secret) {
-	jsonData, _ := json.MarshalIndent(&data, "", " ")
-	mode := int(0660)
-	err := ioutil.WriteFile(".scuttle.json", jsonData, os.FileMode(mode))
-	if err != nil {
-		log.Fatal(err)
-	}
-}
-
-func decode(encoded []byte) ([]byte, error) {
-	// Introduced in v0.7 - we base64 wrap all raw data now, so we have to
-	// attempt to decode. This will return the error object if it fails
-	// and should only be invoked when encoding is set to base64
-	decoded, err := b64.StdEncoding.DecodeString(string(encoded))
-
-	return decoded, err
-}
-
-// encrypt will encrypt the input plaintext with the specified symmetric key
-// example keyName: "projects/PROJECT_ID/locations/global/keyRings/RING_ID/cryptoKeys/KEY_ID"
-func encryptSymmetric(keyName string, plaintext []byte, encoding string) ([]byte, error) {
-	ctx := context.Background()
-	client, err := cloudkms.NewKeyManagementClient(ctx)
-	if err != nil {
-		return nil, err
-	}
-	// as of v0.7 we added base64 encoding by default.
-	// this switcher lets you disable this behavior
-	encoded := string(plaintext)
-	if encoding == "base64" {
-		encoded = b64.StdEncoding.EncodeToString(plaintext)
-	}
-
-	// Build the request.
-	req := &kmspb.EncryptRequest{
-		Name:      keyName,
-		Plaintext: []byte(encoded),
-	}
-	// Call the API.
-	resp, err := client.Encrypt(ctx, req)
-	if err != nil {
-		return nil, err
-	}
-	return resp.Ciphertext, nil
-}
-
-// decrypt will decrypt the input ciphertext bytes using the specified symmetric key
-// example keyName: "projects/PROJECT_ID/locations/global/keyRings/RING_ID/cryptoKeys/KEY_ID"
-func decryptSymmetric(keyName string, ciphertext []byte) ([]byte, error) {
-	ctx := context.Background()
-	client, err := cloudkms.NewKeyManagementClient(ctx)
-	if err != nil {
-		return nil, err
-	}
-
-	// Build the request.
-	req := &kmspb.DecryptRequest{
-		Name:       keyName,
-		Ciphertext: ciphertext,
-	}
-	// Call the API.
-	resp, err := client.Decrypt(ctx, req)
-	if err != nil {
-		// if this fails, its likely network or permissions related
-		return nil, err
-	}
-
-	// return the decrypted data, and the error object
-	return resp.Plaintext, err
-
-}
-
-func userInput() []byte {
-	// Read STDIN (keyboard, interactive) until the user sends a manual EOF
-	// with CTRL+D on WIN keyboards, CMD+D on mac.
-	fmt.Println("Enter the data you want to encrypt. END with CTRL+D or CMD+D")
-	rdr := bufio.NewReader(os.Stdin)
-	var lines []byte
-	for {
-		line, err := rdr.ReadBytes('\n')
-		if err == io.EOF {
-			break
-		}
-		if err != nil {
-			log.Fatalf("Error on input: %s", err)
-		}
-
-		// append scanned input to the array
-		lines = append(lines, line...)
-
-	}
-	return lines
-}
-
-func checkEnv(key string) bool {
-	value := os.Getenv(key)
-	if len(value) == 0 {
-		return false
-	}
-	return true
-}
 
 func main() {
 	app := cli.NewApp()
@@ -199,7 +38,7 @@ func main() {
 			},
 			Action: func(c *cli.Context) error {
 				// Determine if the app is configured
-				if !checkEnv("SCTL_KEY") {
+				if !CheckEnv("SCTL_KEY") {
 					log.Fatal("Missing Env configuration: SCTL_KEY")
 				}
 				var plaintext []byte
@@ -232,24 +71,32 @@ func main() {
 				secret_name := c.Args().First()
 
 				secret_encoding := ""
+				// determine if we need to base64 the raw text, defaults
+				// to true.
 				if c.Bool("no-decode") == true {
+					// skip encoding, encode as plain value
 					secret_encoding = "plain"
 				} else {
+					// encode value as base64 compressed string
 					secret_encoding = "base64"
+					plaintext = []byte(b64Encode(plaintext))
 				}
 
-				cypher, err := encryptSymmetric(c.String("key"), plaintext, secret_encoding)
+				// encryption doesn't care about format, give it everything we've
+				// pre-processed up to this point.
+				cypher, err := encryptSymmetric(c.String("key"), plaintext)
 				if err != nil {
 					log.Fatal(err)
 				}
-				encoded := b64.StdEncoding.EncodeToString(cypher)
+				// re-encode the binary data we got back.
+				encoded := b64Encode(cypher)
 				to_add := Secret{
 					Name:       strings.ToUpper(secret_name),
 					Cyphertext: encoded,
 					Created:    time.Now(),
 					Encoding:   secret_encoding,
 				}
-				addSecret(to_add)
+				AddSecret(to_add)
 				return nil
 			},
 		},
@@ -270,11 +117,11 @@ func main() {
 					log.Fatal("Empty input detected. Aborting")
 				}
 
-				cypher, err := encryptSymmetric(c.String("key"), plaintext, "base64")
+				cypher, err := encryptSymmetric(c.String("key"), plaintext)
 				if err != nil {
 					log.Fatal(err)
 				}
-				encoded := b64.StdEncoding.EncodeToString(cypher)
+				encoded := b64Encode(cypher)
 
 				fmt.Println("Hello, I've shared some data with you with sctl! https://github.com/vapor-ware/sctl")
 				fmt.Println("Once installed, run the following commands to view this sensitive information")
@@ -298,7 +145,7 @@ func main() {
 			},
 			Action: func(c *cli.Context) error {
 				if len(c.Args()) >= 1 {
-					decoded, err := b64.StdEncoding.DecodeString(c.Args().First())
+					decoded, err := b64Decode([]byte(c.Args().First()))
 					if err != nil {
 						log.Fatal(err)
 					}
@@ -318,7 +165,7 @@ func main() {
 			Usage: "rm a secret",
 			Action: func(c *cli.Context) error {
 				secret_name := strings.ToUpper(c.Args().First())
-				rmSecret(secret_name)
+				RmSecret(secret_name)
 				return nil
 			},
 		},
@@ -350,7 +197,7 @@ func main() {
 				},
 			},
 			Action: func(c *cli.Context) error {
-				if !checkEnv("SCTL_KEY") {
+				if !CheckEnv("SCTL_KEY") {
 					log.Fatal("Missing Env configuration: SCTL_KEY")
 				}
 
@@ -362,7 +209,7 @@ func main() {
 				secrets = ReadSecrets()
 				for _, secret := range secrets {
 					// uncan the base64
-					decoded, err := b64.StdEncoding.DecodeString(secret.Cyphertext)
+					decoded, err := b64Decode([]byte(secret.Cyphertext))
 					if err != nil {
 						log.Fatal(err)
 					}
@@ -373,7 +220,7 @@ func main() {
 					}
 					// switch output if encoding == base64
 					if secret.Encoding == "base64" {
-						cypher, err = decode(cypher)
+						cypher, err = b64Decode(cypher)
 						if err != nil {
 							log.Fatal(err)
 						}
@@ -381,7 +228,7 @@ func main() {
 						log.Printf("skipping decode of %v due to encoding != base64", secret.Name)
 					}
 					// Format the decrypted data for ENV consumption
-					skrt := fmt.Sprintf("%s=%s", secret.Name, cypher)
+					skrt := fmt.Sprintf("%s=%v", secret.Name, string(cypher))
 					// Append it to the command exec environment
 					cmd.Env = append(cmd.Env, skrt)
 				}
